+++ .github/workflows/ci.yml
@@ -536,6 +536,25 @@ jobs:
         run: |
           ./scripts/run-rustls-integration.sh
 
+  semver-checks:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+        with:
+          submodules: 'recursive'
+      - name: Check semver (Default Features)
+        uses: obi1kenobi/cargo-semver-checks-action@v2
+        with:
+          package: aws-lc-rs
+          feature-group: default-features
+      - name: Check semver (FIPS)
+        uses: obi1kenobi/cargo-semver-checks-action@v2
+        with:
+          package: aws-lc-rs
+          feature-group: only-explicit-features
+          features: fips
+
   copyright:
     runs-on: ubuntu-latest
     steps:
